테이블 수정. ItemUsingAction에 모션 추가

+++ Assets/Scripts/Inventory/Inventory/Inventory.cs
@@ -71,6 +71,7 @@ public class Inventory : IInventory
         List<IItem> itemList = GetCorrectTypeItemList(deleteItem.GetData().itemType);
         for (int i = 0; i < itemList.Count; i++)
         {
+            if (itemList[i] is null) continue;
             if (itemList[i].GetData().id == deleteItem.GetData().id)
             {
                 itemList[i] = null;
@@ -83,9 +84,7 @@ public class Inventory : IInventory
     public void SwapItem(ItemType type, int start, int end)
     {
         List<IItem> itemList = GetCorrectTypeItemList(type);
-        IItem tmpItem = itemList[end];
-        itemList[end] = itemList[start];
-        itemList[start] = tmpItem;
+        (itemList[end], itemList[start]) = (itemList[start], itemList[end]);
     }
     public void EqipItem(ItemType type, int index)
     {
@@ -109,7 +108,7 @@ public class Inventory : IInventory
 
         if (itemList[index].IsUsable())
         {
-            Debug.Log("selecet");
+            Debug.Log("select");
             player.SelectItem(itemList[index]);
         }
     }
